+ CHG: Fix some issues found with typing annotation for Numpy 1.2

+++ pyxel/data_structure/photon.py
@@ -25,7 +25,7 @@ class Photon(Array):
     """
 
     # TODO: add unit (ph)
-    EXP_TYPE = np.int
+    EXP_TYPE = int
     TYPE_LIST = (
         np.int32,
         np.int64,
